requirements.txt
numpy
cv
uvicorn
fastapi
"fastapi[standard]"
pydantic
tensorflow 
pillow